requirements.txt
flask
pandas
python-dotenv
mysql-connector-python
openpyxl
Flask-SocketIO
